requirements.txt
Flask>=3.0,<4.0
gunicorn>=21.2,<23.0
openai>=1.50.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0


